Rewrite of logging functionality (no more conditionalLogging method)
I studied the python logging docs and implemented logging in the
recommended way now. A logger instance can be given to the
CrabController instance if it exists and log messages are appended
to a new logger instance with the given logger as the parent instance.
This allows to notice that the logs messages where raised in the
crabFunctions library. A new logger is added  if no logger is given as a argument.
Furthermore the function checks if the root logger is configured (by checking if a
handler is present) and sets some basic logging configuration otherwise.