(HistClass_update) Included first updated versions of the WriteAll function
To have additional functionality two additional versions of the WriteAll function
are included. They get an additional string parameter were different substrings
are that have to be in the histogram name. The third version has an additional
string parameter were different substrings that should be vetoed in the histogram
name.