(HistClass_modification) Completed the dunctionallity of the WriteAll functions
The WriteAll functions were completed, it is now also possible to give a string
of names (seperated by a ':') that should be contained in the histo name. Also
a version were the contains requirement is expanded with a veto requirement is
included. Also an example how to use this functionalliyt in your analysis was
included.