Included check, whether the gridpack exists on dcache. The user has the chance to overwrite it then. Otherwise the program crashes anyways.
Changed the createAndUploadGridpack function. The required parameter is
now a list (even if it has only one entry). This way, several files with
relative paths can be added to the grid pack without problems from tar.
If no list is used, the given string will be split up in its characters
and appended

changed the ls command in the prologue to ls -R (lists also the subdir
contents)