aix3adb: added search feature
skims can now be searched with the two functions
searchMCSkimsAndSamples
searchDataSkimsAndSamples
they require two dicts as parameters, the first for the skim criteria
the second for the sample criteria. each criterion is defined by a
key, value pair of the dict. The key is allowed to start with a
comparator < < >= <= to compare a number. percent signs should work
as wildcards.