cesubmit.py make position of executable flexible
Changed way in which the executable is given to the jdl-script.
Now it is possible to use executables which are not located in the
head directory of the gridpackage by giving not only the name but
also the relative path to the executable in the config file.
For example: executable=EventClassFactory/ce_wrapper.sh