Temporarily remove import of log level
This commit will remove the import of the log level. This feature has been
explicitly requested by our workgroup, but has apparently been neglected by the
CRAB authors and does now crash on importing. Consequently, this commit will
disable it until we have further clarification from the CRAB authors on the
future of this feature.