aix3adb: do not require kinit if available
This commit uses the klist command to check if a kerberos
certificate exists and may be used to obtain a CERN SSO cookie.
This avoids entering the CERN password every time at startup
in each script that communicates with aix3adb.

Users may renew it or proceed by pressing enter.