crabFunctions: Task add inDB property and updateFromDB
This commti add the property inDB to the CrabTask function.
This property uses the new function updateFromDB. Which updated
the Skim and Sample member objects of the class.

A crabtask  is considered inDB if it has a sampole / skim with the same:
* sample name
* skimmer_version
* globalTag
* (for data) json file

The query in updateFromDB is build based on those criteria.